% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stow-package.R
\docType{package}
\name{stow-package}
\alias{stow-package}
\title{stow: Durable Package-Scoped File Downloads and Caching}
\description{
The stow package downloads files into durable, package-scoped user data
directories. It supports ETag-aware names, offline lookup, validation, and
transactional replacement without processing or publishing artifacts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cole-brokamp/stow}
  \item Report bugs at \url{https://github.com/cole-brokamp/stow/issues}
}

}
\author{
\strong{Maintainer}: Cole Brokamp \email{cole@colebrokamp.com} (\href{https://orcid.org/0000-0002-0289-3151}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Cole Brokamp \email{cole@colebrokamp.com} (\href{https://orcid.org/0000-0002-0289-3151}{ORCID}) [copyright holder]
}

}
\keyword{internal}
